Bugfix `Leaf` pathing
All code was mistakenly using the label of a `Leaf` to determine its
path. This is obviously a mistake as the label of a `Leaf` is affected
by the value stored in that `Leaf`.

This commit also refactors the `label` function inside of `Store`. We
will use this in a future commit to improve map identification in
`Store`s.